+++ vendas/src/main/java/com/example/vendas/rest/Controller/DTO/InfitemPedidoDTO.java
@@ -4,11 +4,13 @@ package com.example.vendas.rest.Controller.DTO;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class InfitemPedidoDTO {
     private String descricao;
-    private float preco;
-    private int quantidade;
+    private BigDecimal preco;
+    private Integer quantidade;
 
 }
